pyproject.toml
[tool.pdm]
distribution = true

[tool.pdm.build]
includes = [
    "nonebot_plugin_genshin_cos",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "nonebot-plugin-genshin-cos"
version = "0.3.5"
description = "米游社原神cos图获取"
authors = [
    {name = "Cvandia",email = "106718176+Cvandia@users.noreply.github.com"},
]
dependencies = [
    "nonebot-plugin-apscheduler>=0.1.3",
    "nonebot-adapter-onebot>=2.2.0",
    "nonebot2>=2.2.0",
    "httpx>=0.19.0",
    "aiofiles>=0.7.0",
    "pathlib>=1.0.1",
    "ujson>=5.5.0",
]
requires-python = "<4,>=3.9"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
dev = ["pre-commit>=4.0.1"]
